fix(code-block): restore fallback sizing and bundle boundary
