+++ src/components/CodeBlockNode/CodeBlockNode.vue
@@ -13,12 +13,10 @@ import { MARKSTREAM_NODE_LIFECYCLE_KEY } from '../../utils/nodeLifecycle'
 import { resolveLanguageIcon } from '../../utils/resolveLanguageIcon'
 import { safeCancelRaf, safeRaf } from '../../utils/safeRaf'
 import PreCodeBlock from '../PreCodeNode/PreCodeBlock.vue'
+import { preCodeThemeLooksDark, resolvePreCodeThemeName, resolvePreCodeThemePalette } from '../PreCodeNode/preCodeTheme'
 import {
   DEFAULT_PRE_CODE_FONT_SIZE,
   DEFAULT_PRE_CODE_LINE_HEIGHT,
-  preCodeThemeLooksDark,
-  resolvePreCodeThemeName,
-  resolvePreCodeThemePalette,
   resolvePreCodeVisualOptions,
 } from '../PreCodeNode/preCodeVisual'
 import { isDiffCodeBlock, resolveCodeBlockHeader, resolveDiffHideUnchangedRegionsOption } from './codeBlockHeader'
