+++ test/node-renderer-measurement-performance.test.ts
@@ -283,6 +283,58 @@ describe('node renderer measurement performance', () => {
     wrapper.unmount()
   })
 
+  it('passes estimated content height to pre blocks without duplicating header height', async () => {
+    vi.spyOn(HTMLElement.prototype, 'clientWidth', 'get').mockImplementation(() => 320)
+
+    const NodeRenderer = (await import('../src/components/NodeRenderer')).default
+    const nodes = [
+      createCodeBlock(1),
+      {
+        ...createCodeBlock(2),
+        code: Array.from({ length: 80 }, (_, index) => `console.log(${index})`).join('\n'),
+      },
+    ]
+    const wrapper = mount(NodeRenderer, {
+      props: {
+        nodes,
+        renderCodeBlocksAsPre: true,
+        viewportPriority: false,
+        codeBlockProps: { showHeader: true },
+        virtualScroll: {
+          enabled: true,
+          sessionKey: 'estimated-pre-content-height',
+        },
+      },
+    })
+
+    await flushAll()
+    const state = setupState(wrapper)
+    const readItems = () => Array.isArray(state.renderedItems)
+      ? state.renderedItems
+      : state.renderedItems.value
+    const readEstimates = () => Array.isArray(state.estimatedNodeHeights)
+      ? state.estimatedNodeHeights
+      : state.estimatedNodeHeights.value
+
+    for (const [index, item] of readItems().entries()) {
+      const estimate = readEstimates()[index]
+      expect(estimate.kind).toBe('code-block')
+      expect(estimate.height - estimate.contentHeight).toBe(37)
+      expect(item.bindings.reservedHeightPx).toBe(estimate.contentHeight)
+    }
+
+    await wrapper.setProps({ codeBlockProps: { showHeader: false } })
+    await flushVueOnly()
+
+    for (const [index, item] of readItems().entries()) {
+      const estimate = readEstimates()[index]
+      expect(estimate.height).toBe(estimate.contentHeight)
+      expect(item.bindings.reservedHeightPx).toBe(estimate.contentHeight)
+    }
+
+    wrapper.unmount()
+  })
+
   it('recomputes estimated heights when custom code block components change', async () => {
     vi.spyOn(HTMLElement.prototype, 'clientWidth', 'get').mockImplementation(() => 640)
