+++ test/pre-code-block-shell.test.ts
@@ -1,7 +1,7 @@
 import { mount } from '@vue/test-utils'
 import { describe, expect, it } from 'vitest'
 import PreCodeBlock from '../src/components/PreCodeNode/PreCodeBlock.vue'
-import { resolvePreCodeThemePalette } from '../src/components/PreCodeNode/preCodeVisual'
+import { resolvePreCodeThemePalette } from '../src/components/PreCodeNode/preCodeTheme'
 
 const node = {
   type: 'code_block' as const,
